+++ app/constants/translations/de/common.js
@@ -1,4 +1,25 @@
 export default {
-  hello: 'Hallo',
   languageSelector: 'Wähle deine Sprache',
+  description: {
+    description:
+      'Das Cushing-Syndrom tritt auf, wenn der Körper ' +
+      'mit der Zeit zu viel des Hormons Cortisol hat. ' +
+      'Es kann das Ergebnis der oralen Einnahme von ' +
+      'Kortikosteroiden oder der Körperproduktion ' +
+      'von zu viel Cortisol sein.',
+  },
+  symptom: {
+    description: 'Symptome',
+    items: [
+      'Gewichtszunahme und Fettgewebsablagerungen, ' +
+        'insbesondere um die Körpermitte und den oberen ' +
+        'Rücken, im Gesicht (Mondgesicht) und zwischen ' +
+        'den Schultern (Büffelhals)',
+      'Rosa oder violette Striae gravidarum auf der Haut ' +
+        ' von Bauch, Oberschenkeln, Brüsten und Armen.',
+      'Dünne und empfindliche Haut, die zu Blutergüssen neigt',
+      'Langsame Heilung von Wunden, Insektenstichen und Infektionen',
+      'Akne',
+    ],
+  },
 };