+++ app/constants/translations/en/common.js
@@ -1,4 +1,25 @@
 export default {
-  hello: 'Hello',
   languageSelector: 'Select your Language',
+  description: {
+    description:
+      "Cushing's syndrome occurs when the body " +
+      'has too much of the hormone cortisol over ' +
+      'time. It can be the result of taking ' +
+      'corticosteroids by mouth or the body making ' +
+      'too much cortisol.',
+  },
+  symptom: {
+    description: 'Symptoms',
+    items: [
+      'Weight gain and fatty tissue deposits, particularly ' +
+        'around the midsection and upper back, in the face ' +
+        '(moon face), and between the shoulders (buffalo ' +
+        'neck)',
+      'Pink or purple striae gravidarum on the skin of ' +
+        'the abdomen, thighs, breasts, and arms.',
+      'Thin and fragile skin, prone to bruising',
+      'Slow healing of wounds, insect bites and infections',
+      'Acne',
+    ],
+  },
 };