+++ app/constants/translations/en/navigate.js
@@ -1,32 +1,6 @@
 export default {
-  home: {
-    title: 'Home',
-  },
-  setting: {
-    title: 'Setting',
-  },
-  description: {
-    title: 'Description',
-    description:
-      "Cushing's syndrome occurs when the body " +
-      'has too much of the hormone cortisol over ' +
-      'time. It can be the result of taking ' +
-      'corticosteroids by mouth or the body making ' +
-      'too much cortisol.',
-  },
-  symptom: {
-    title: 'Symptoms',
-    description: 'Symptoms',
-    items: [
-      'Weight gain and fatty tissue deposits, particularly ' +
-        'around the midsection and upper back, in the face ' +
-        '(moon face), and between the shoulders (buffalo ' +
-        'neck)',
-      'Pink or purple striae gravidarum on the skin of ' +
-        'the abdomen, thighs, breasts, and arms.',
-      'Thin and fragile skin, prone to bruising',
-      'Slow healing of wounds, insect bites and infections',
-      'Acne',
-    ],
-  },
+  home: 'Home',
+  setting: 'Setting',
+  description: 'Description',
+  symptom: 'Symptoms',
 };