+++ app/constants/translations/es/common.js
@@ -1,4 +1,27 @@
 export default {
-  hello: 'Hola',
   languageSelector: 'Elige tu idioma',
+  description: {
+    description:
+      'El síndrome de Cushing aparece cuando ' +
+      'el cuerpo tiene demasiada hormona cortisol a lo largo ' +
+      'del tiempo. Puede ser el resultado de tomar ' +
+      'corticosteroides por vía oral o de que el' +
+      ' cuerpo produzca demasiado cortisol.',
+  },
+  symptom: {
+    description: 'Síntomas',
+    items: [
+      'Aumento de peso y depósitos de tejido adiposo, ' +
+        'particularmente alrededor de la sección media ' +
+        'del cuerpo y la parte superior de la espalda, ' +
+        'en la cara (cara de luna llena) y entre los ' +
+        'hombros (cuello de bisonte)',
+      'Estrías gravídicas color rosa o púrpura en la ' +
+        'piel del abdomen, muslos, mamas y brazos.',
+      'Piel fina y frágil, propensa a hematomas',
+      'Cicatrización lenta de heridas, picaduras de ' +
+        'insectos e infecciones',
+      'Acné',
+    ],
+  },
 };